Fix compilation problems with some compilers that expect isinf() to accept only floats.

+++ src/Platforms/Gcc/UtestPlatform.cpp
@@ -239,7 +239,7 @@ static int IsNanImplementation(double d)
 
 static int IsInfImplementation(double d)
 {
-    return isinf(d);
+    return isinf((float)d);
 }
 
 double (*PlatformSpecificFabs)(double) = fabs;
